chartOptiosBuilder-commonCOnfiguration moved constant
