+++ src/components/visualizations/chart/legend/helpers.ts
@@ -9,9 +9,9 @@ import inRange = require("lodash/inRange");
 import { IHeatmapLegendItem } from "../../typings/legend";
 import { LEFT, RIGHT, TOP, BOTTOM } from "./PositionTypes";
 import { formatLegendLabel, isAreaChart, isOneOfTypes, isTreemap } from "../../utils/common";
-import { supportedDualAxesChartTypes } from "../chartOptionsBuilder";
 import { ISeriesItem } from "../../../../interfaces/Config";
 import { VisualizationTypes } from "../../../../constants/visualizationTypes";
+import {supportedDualAxesChartTypes} from "../../../../constants/chart";
 
 export const RESPONSIVE_ITEM_MIN_WIDTH = 200;
 export const RESPONSIVE_VISIBLE_ROWS = 2;
