+++ src/components/visualizations/chart/legend/legendBuilder.ts
@@ -13,8 +13,8 @@ import {
 } from "../../utils/common";
 import { VisualizationTypes } from "../../../../constants/visualizationTypes";
 import { ILegendOptions, LegendOptionsItemType, DEFAULT_LEGEND_CONFIG } from "../../typings/legend";
-import { supportedDualAxesChartTypes } from "../chartOptionsBuilder";
 import { isStackedChart } from "./helpers";
+import {supportedDualAxesChartTypes} from "../../../../constants/chart";
 
 function isHeatmapWithMultipleValues(chartOptions: any) {
     const { type } = chartOptions;
