+++ src/constants/chart.ts
@@ -1,0 +1,11 @@
+// (C) 2019 GoodData Corporation
+import {VisualizationTypes} from "./visualizationTypes";
+
+export const supportedDualAxesChartTypes = [
+    VisualizationTypes.COLUMN,
+    VisualizationTypes.BAR,
+    VisualizationTypes.LINE,
+    VisualizationTypes.AREA,
+    VisualizationTypes.COMBO,
+    VisualizationTypes.COMBO2,
+];
